requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
Django==4.1.7
django-bootstrap4==22.3
django-phonenumber-field==7.0.2
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
fonttools==4.39.2
future==0.18.2
html5lib==1.1
phonenumbers==8.13.7
Pillow==9.4.0
pipdeptree==2.13.0
pycparser==2.21
pydyf==0.5.0
pyphen==0.14.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tinycss2==1.2.1
weasyprint==58.1
webencodings==0.5.1
zopfli==0.2.2